aria-app/client/requirements.txt
# ARIA Edge Device Client Requirements
# For Raspberry Pi deployment

# Core dependencies
requests>=2.31.0

# Face Recognition
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
tensorflow>=2.15.0
keras>=2.15.0
keras-facenet>=0.3.0

# Hardware (Raspberry Pi only)
RPi.GPIO>=0.7.1; platform_machine == "armv7l" or platform_machine == "aarch64"

# Optional: for development/testing
python-dotenv>=1.0.0

